правки после 1-го ревью

+++ yatube/posts/views.py
@@ -1,9 +1,15 @@
-from django.shortcuts import get_object_or_404, render, redirect
 from django.contrib.auth.decorators import login_required
+from django.shortcuts import get_object_or_404, redirect, render
 
-from .models import Group, Post, User, Follow
 from .forms import CommentForm, PostForm
+from .models import Follow, Group, Post, User
 from .page_separation import get_page_obj
+
+
+def paginator(request, posts):
+    page_number = request.GET.get('page')
+    page_obj = get_page_obj(posts, page_number)
+    return page_obj
 
 
 def index(request):
@@ -12,10 +18,8 @@ def index(request):
         'group'
     )
     template = 'posts/index.html'
-    page_number = request.GET.get('page')
-    page_obj = get_page_obj(posts, page_number)
+    page_obj = paginator(request, posts)
     context = {
-        'posts': posts,
         'page_obj': page_obj,
     }
     return render(request, template, context)
@@ -24,12 +28,10 @@ def index(request):
 def group_posts(request, slug):
     group = get_object_or_404(Group, slug=slug)
     posts = group.posts.select_related('author')
-    page_number = request.GET.get('page')
-    page_obj = get_page_obj(posts, page_number)
+    page_obj = paginator(request, posts)
     template = 'posts/group_list.html'
     context = {
         'group': group,
-        'posts': posts,
         'page_obj': page_obj,
     }
     return render(request, template, context)
@@ -38,20 +40,17 @@ def group_posts(request, slug):
 def profile(request, username):
     author = get_object_or_404(User, username=username)
     posts = author.posts.select_related('group')
-    posts_count = posts.count()
-    page_number = request.GET.get('page')
-    page_obj = get_page_obj(posts, page_number)
+    page_obj = paginator(request, posts)
     template = 'posts/profile.html'
-    following = (request.user != author
-                 and request.user.is_authenticated
-                 and Follow.objects.filter(
-                     user=request.user,
-                     author=author
-                 ).exists())
+    following = (
+        request.user.is_authenticated
+        and Follow.objects.filter(
+            user=request.user,
+            author=author
+        ).exists())
     context = {
         'author': author,
         'page_obj': page_obj,
-        'posts_count': posts_count,
         'following': following,
     }
     return render(request, template, context)
@@ -59,7 +58,6 @@ def profile(request, username):
 
 def post_detail(request, post_id):
     post = get_object_or_404(Post, pk=post_id)
-    author_posts_count = post.author.posts.count()
     template = 'posts/post_detail.html'
     comments = post.comments.all()
     form = CommentForm()
@@ -71,8 +69,6 @@ def post_detail(request, post_id):
                  ).exists())
     context = {
         'post': post,
-        'post_id': post_id,
-        'author_posts_count': author_posts_count,
         'form': form,
         'comments': comments,
         'following': following,
@@ -131,8 +127,7 @@ def add_comment(request, post_id):
 def follow_index(request):
     username = request.user
     posts = Post.objects.filter(author__following__user=username)
-    page_number = request.GET.get('page')
-    page_obj = get_page_obj(posts, page_number)
+    page_obj = paginator(request, posts)
     template = 'posts/follow.html'
     context = {
         'page_obj': page_obj,
@@ -144,7 +139,6 @@ def follow_index(request):
 def profile_follow(request, username):
     author = get_object_or_404(User, username=username)
     if username != request.user.username:
-        # author = get_object_or_404(User, username=username)
         Follow.objects.get_or_create(
             user=request.user,
             author=author
@@ -154,9 +148,6 @@ def profile_follow(request, username):
 
 @login_required
 def profile_unfollow(request, username):
-    get_object_or_404(
-        Follow,
-        user=request.user,
-        author__username=username
-    ).delete()
+    author = get_object_or_404(User, username=username)
+    Follow.objects.filter(user=request.user, author=author).delete()
     return redirect('posts:profile', username=username)
